Inference Gateway в полностью готовой версии, Inference почти доделан - осталось добавить поддержу другой модели
